+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-beta.83
+
+### Patch Changes
+
+- Add local backup for action effects
+
+## 0.1.1-beta.82
+
+### Patch Changes
+
+- Fix preview server missing skill action
+
+## 0.1.1-beta.81
+
+### Patch Changes
+
+- Add local action when opened in non-editor env
+
 ## 0.1.1-beta.80
 
 ### Patch Changes